[migrate] replace old Lark class with new Lark App [add] Parcel 2 bundler configuration
