+++ .gitignore
@@ -2,6 +2,7 @@ node_modules/
 package-lock.json
 yarn.lock
 yarn-error.log
+.parcel-cache/
 dist/
 docs/
 .env
