+++ src/index.ts
@@ -4,6 +4,5 @@ globalThis.fetch ||= fetch;
 
 export * from './type';
 export * from './Lark';
-export * from './LarkApp';
-export * from './module';
+// export * from './module';
 export * from './model/BiDataTable';
