+++ src/module/base.ts
@@ -1,11 +1,11 @@
-import { Lark } from '../Lark';
+import { LarkApp } from '../Lark';
 
 export abstract class LarkModule {
-    core: Lark;
+    core: LarkApp;
     id?: string;
     abstract baseURI: string;
 
-    constructor(core: Lark, id?: string) {
+    constructor(core: LarkApp, id?: string) {
         this.core = core;
         this.id = id;
     }
