+++ tsconfig.json
@@ -5,7 +5,6 @@
         "module": "CommonJS",
         "moduleResolution": "Node",
         "experimentalDecorators": true,
-        "importHelpers": true,
         "declaration": true,
         "skipLibCheck": true,
         "lib": ["ES2021", "DOM"],
